Move EWMA into its own module and turn it into a transformer + Unit.
